Dir: fix conversion warning when compiling with clang
* Add class name and use it when printing debug
* get_node(path, root, root_len): change `root_len` type from `int` to
  `size_t`

+++ Dir.hh
@@ -37,7 +37,7 @@ public:
 	int get_parent_path(Buffer *path, DirNode *ls, int depth = 1);
 	int get_list(DirNode* list, const char *path, int depth = -1);
 	int get_symlink(DirNode* list);
-	DirNode* get_node(Buffer* path, const char* root, int root_len);
+	DirNode* get_node(Buffer* path, const char* root, size_t root_len);
 
 	int refresh_by_path(Buffer* path);
 	void dump();
